+++ submission/.gitignore
@@ -1,1 +1,2 @@
 server_config.py
+**/__pycache__/*
