+++ submission/lib/test_trace_build.py
@@ -1,46 +1,10 @@
-from .data_types import Trace, BusinessIndex, PlatformIndex
-import numpy as np
-
-from collections import defaultdict
-
 import pandas as pd
 import time
 import sys
 import random
 
-def parse(traces):
-    """
-    Input:
-        traces - list of Trace objects as in main.py
-
-    Output:
-        dictionary {<trace_id> : [<processed traces>]}
-
-    Processing includes sorting and removing the time of the children from the elapsed time of the parent
-    """
-    
-    trace_dict = defaultdict(list) # { trace_id : [traces]}
-
-    # Split traces according to their id
-    for trace in traces:
-        trace_dict[trace.trace_id].append(trace)
-
-    # Sort them according to start time
-    for trace in trace_dict.values():
-        trace.sort(key=lambda x: x.start_time)
-
-    # Remove time of the children from the parent
-    for trace in trace_dict.values():
-        times = defaultdict(int)
-        
-        for i in range(len(trace) - 1, 0, -1):
-            times[trace[i].pid] += trace[i].elapsed_time
-        
-        for el in trace:
-            el.elapsed_time -= times[el.id]
-
-    return trace_dict
-
+from utils.data_types import *
+from utils.trace_build import *
 
 if __name__ == '__main__':
     print(f'Testing {__file__}')
