+++ submission/main.py
@@ -9,9 +9,9 @@ from collections import deque
 from kafka import KafkaConsumer
 
 from server_config import SERVER_CONFIGURATION
-from utils.data_types import PlatformIndex, BusinessIndex, Trace
+from lib.utils.data_types import PlatformIndex, BusinessIndex, Trace
 
-from utils import trace_build
+from lib.utils import trace_build
 
 # Three topics are available: platform-index, business-index, trace.
 # Subscribe at least one of them.
